Added files used for deriving the resting state time series
Region time series were extracted and analysed when 100% data was available (i.e., if one participant was missing BOLD value data in more tan 80% of the voxels in a region, that region was excluded from the analysis). These scripts allow for checking of NaN values and removing the regions that showed NaN values.